+++ local-addons/merctrans_projects/models/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
 # from . import merctrans_sales
-from . import merctrans_clients, merctrans_pos, merctrans_projects
+from . import merctrans_clients, merctrans_pos, merctrans_projects, merctrans_invoices
