+++ local-addons/merctrans_projects/views/projects_menus.xml
@@ -29,6 +29,15 @@
     <menuitem name="Purchase Order" id="po_menu_in_base_menu" parent="project_base_menu" action="merctrans_pos_actions" sequence="4" />
 
     <!-- invoices tree view -->
+    <!-- Auto increase invoice_id-->
+    <record id="increment_invoices_id" model="ir.sequence">
+      <field name="name">ID Incrementation</field>
+      <field name="code">increment_invoice_id</field>
+      <field name="prefix"></field>
+      <field name="padding">3</field>
+      <field name="company_id" eval="False" />
+    </record>
+
     <record id="invoices_view_tree" model="ir.ui.view">
       <field name="name">invoices tree</field>
       <field name="model">merctrans.invoices</field>
@@ -121,7 +130,7 @@
                 <field name="project_manager" />
                 <field name="project_id" />
                 <field name="project_name" />
-                <field name="client" options="{'no_create': True, 'no_create_edit':True}"/>
+                <field name="client" options="{'no_create': True, 'no_create_edit':True}" />
                 <field name="services_ids" widget="many2many_tags" />
                 <field name="source_language" />
                 <field name="target_language" />
@@ -131,13 +140,13 @@
               </group>
               <group>
                 <field name="currency_id" />
-                <field name="discount" widget="percentage"/>
+                <field name="discount" widget="percentage" />
                 <field name="work_unit" />
                 <field name="volume" />
                 <field name="sale_rate" />
                 <field name="project_value" />
-                <field name="total_po_value"/>
-                <field name="project_margin" widget="percentage"/>
+                <field name="total_po_value" />
+                <field name="project_margin" widget="percentage" />
                 <field name="project_status" />
                 <field name="payment_status" />
               </group>
@@ -152,9 +161,9 @@
                     <field name="work_unit" />
                     <field name="volume" />
                     <field name="sale_rate_per_work_unit" />
-                    <field name="due_date" require="True"/>
+                    <field name="due_date" require="True" />
                     <field name="currency_id" />
-                    <field name="po_value"/>
+                    <field name="po_value" />
                     <field name="po_status" />
                     <field name="payment_status" />
                   </tree>
@@ -168,13 +177,13 @@
                         <field name="service" />
                         <field name="valid_date" />
                         <field name="start_date" require="True" />
-                        <field name="due_date"  />
+                        <field name="due_date" />
                       </group>
                       <group>
                         <field name="work_unit" />
                         <field name="volume" />
                         <field name="sale_rate_per_work_unit" />
-                        <field name="po_value"/>
+                        <field name="po_value" />
                         <field name="po_status" />
                         <field name="payment_status" />
                       </group>
@@ -201,9 +210,9 @@
         <tree>
           <field name="project_id" />
           <field name="project_name" />
-          <field name="client" widget="many2onebutton"/>
+          <field name="client" widget="many2onebutton" />
           <!-- <field name="client_name" /> -->
-          <field name="services_ids" widget="many2many_tags"/>
+          <field name="services_ids" widget="many2many_tags" />
           <field name="work_unit" />
           <!--          <field name="source_language" />-->
           <!--          <field name="target_language" />-->
@@ -211,8 +220,8 @@
           <field name="sale_rate" />
           <field name="currency_id" />
           <field name="project_value" />
-          <field name="project_margin" widget="percentage"/>
-<!--          <field name="total_po_value"/>-->
+          <field name="project_margin" widget="percentage" />
+          <!--          <field name="total_po_value"/>-->
           <!--          <field name="discount" />-->
           <!--          <field name="start_date" />-->
           <field name="due_date" />
